+++ tests/testthat/test-run-model.R
@@ -20,7 +20,7 @@ test_that("model can be run", {
   output <- readRDS(model_run$output_path)
   expect_equal(colnames(output),
                c("area_level", "area_level_label", "area_id", "area_name",
-                 "sex", "age_group", "age_group_id", "age_group_label",
+                 "sex", "age_group", "age_group_label",
                  "calendar_quarter", "quarter_id", "quarter_label",
                  "indicator", "indicator_id", "indicator_label",
                  "mean", "se", "median", "mode", "lower", "upper"))
@@ -73,7 +73,7 @@ test_that("model can be run", {
   )
 
   ## Check coarse age outputs saved in summar_path
-  coarse_ages <_0 c("Y015_049", "Y015_064", "Y015_999", "Y050_999", "Y000_999", "Y000_064", "Y000_014",
+  coarse_ages <- c("Y015_049", "Y015_064", "Y015_999", "Y050_999", "Y000_999", "Y000_064", "Y000_014",
                    "Y015_024", "Y025_034", "Y035_049", "Y050_064", "Y065_999")
   coarse_age_outputs <- read_output_package(model_run$coarse_output_path)
   expect_setequal(coarse_age_outputs$meta_age_group$age_group, coarse_ages)
@@ -132,7 +132,7 @@ test_that("model can be run without programme data", {
   output <- readRDS(model_run$output_path)
   expect_equal(colnames(output),
                c("area_level", "area_level_label", "area_id", "area_name",
-                 "sex", "age_group", "age_group_id", "age_group_label",
+                 "sex", "age_group", "age_group_label",
                  "calendar_quarter", "quarter_id", "quarter_label",
                  "indicator", "indicator_id", "indicator_label",
                  "mean", "se", "median", "mode", "lower", "upper"))