+++ NEWS.md
@@ -3,6 +3,7 @@
 Version 2.0 established for 2021 UNAIDS estimates. Changes are not guaranteed to be backward compatible with 2020 version of Naomi model.
 
 * Update age group codes to `YXXX_XXX`.
+* Chage `age_group_label` from 00-04 to 0-4 and 05-09 to 5-9.
 * Remove `age_group_id` column.
 
 # naomi 1.0.8